Move a stat test
