+++ sql/core/src/test/scala/org/apache/spark/sql/DataFrameSuite.scala
@@ -34,10 +34,10 @@ import org.apache.spark.api.python.PythonEvalType
 import org.apache.spark.sql.catalyst.{InternalRow, TableIdentifier}
 import org.apache.spark.sql.catalyst.analysis.MultiInstanceRelation
 import org.apache.spark.sql.catalyst.encoders.ExpressionEncoder
-import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeMap, AttributeReference, Cast, EqualTo, ExpressionSet, GreaterThan, Literal, PythonUDF, ScalarSubquery, Uuid}
+import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeReference, Cast, EqualTo, ExpressionSet, GreaterThan, Literal, PythonUDF, ScalarSubquery, Uuid}
 import org.apache.spark.sql.catalyst.optimizer.ConvertToLocalRelation
 import org.apache.spark.sql.catalyst.parser.ParseException
-import org.apache.spark.sql.catalyst.plans.logical.{ColumnStat, Filter, LeafNode, LocalRelation, LogicalPlan, OneRowRelation, Statistics}
+import org.apache.spark.sql.catalyst.plans.logical.{Filter, LeafNode, LocalRelation, LogicalPlan, OneRowRelation}
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
 import org.apache.spark.sql.connector.FakeV2Provider
 import org.apache.spark.sql.execution.{FilterExec, LogicalRDD, QueryExecution, SortExec, WholeStageCodegenExec}
@@ -1897,82 +1897,6 @@ class DataFrameSuite extends QueryTest
     }
   }
 
-  test("SPARK-39834: build the stats for LogicalRDD based on origin stats") {
-    def buildExpectedColumnStats(attrs: Seq[Attribute]): AttributeMap[ColumnStat] = {
-      AttributeMap(
-        attrs.map {
-          case attr if attr.dataType == BooleanType =>
-            attr -> ColumnStat(
-              distinctCount = Some(2),
-              min = Some(false),
-              max = Some(true),
-              nullCount = Some(0),
-              avgLen = Some(1),
-              maxLen = Some(1))
-
-          case attr if attr.dataType == ByteType =>
-            attr -> ColumnStat(
-              distinctCount = Some(2),
-              min = Some(1),
-              max = Some(2),
-              nullCount = Some(0),
-              avgLen = Some(1),
-              maxLen = Some(1))
-
-          case attr => attr -> ColumnStat()
-        }
-      )
-    }
-
-    val outputList = Seq(
-      AttributeReference("cbool", BooleanType)(),
-      AttributeReference("cbyte", ByteType)(),
-      AttributeReference("cint", IntegerType)()
-    )
-
-    val expectedSize = 16
-    val statsPlan = OutputListAwareStatsTestPlan(
-      outputList = outputList,
-      rowCount = 2,
-      size = Some(expectedSize))
-
-    withSQLConf(SQLConf.CBO_ENABLED.key -> "true") {
-      val df = Dataset.ofRows(spark, statsPlan)
-        // add some map-like operations which optimizer will optimize away, and make a divergence
-        // for output between logical plan and optimized plan
-        // logical plan
-        // Project [cb#6 AS cbool#12, cby#7 AS cbyte#13, ci#8 AS cint#14]
-        // +- Project [cbool#0 AS cb#6, cbyte#1 AS cby#7, cint#2 AS ci#8]
-        //    +- OutputListAwareStatsTestPlan [cbool#0, cbyte#1, cint#2], 2, 16
-        // optimized plan
-        // OutputListAwareStatsTestPlan [cbool#0, cbyte#1, cint#2], 2, 16
-        .selectExpr("cbool AS cb", "cbyte AS cby", "cint AS ci")
-        .selectExpr("cb AS cbool", "cby AS cbyte", "ci AS cint")
-
-      // We can't leverage LogicalRDD.fromDataset here, since it triggers physical planning and
-      // there is no matching physical node for OutputListAwareStatsTestPlan.
-      val optimizedPlan = df.queryExecution.optimizedPlan
-      val rewrite = LogicalRDD.buildOutputAssocForRewrite(optimizedPlan.output,
-        df.logicalPlan.output)
-      val logicalRDD = LogicalRDD(
-        df.logicalPlan.output, spark.sparkContext.emptyRDD[InternalRow], isStreaming = true)(
-        spark, Some(LogicalRDD.rewriteStatistics(optimizedPlan.stats, rewrite.get)), None)
-
-      val stats = logicalRDD.computeStats()
-      val expectedStats = Statistics(sizeInBytes = expectedSize, rowCount = Some(2),
-        attributeStats = buildExpectedColumnStats(logicalRDD.output))
-      assert(stats === expectedStats)
-
-      // This method re-issues expression IDs for all outputs. We expect column stats to be
-      // reflected as well.
-      val newLogicalRDD = logicalRDD.newInstance()
-      val newStats = newLogicalRDD.computeStats()
-      val newExpectedStats = Statistics(sizeInBytes = expectedSize, rowCount = Some(2),
-        attributeStats = buildExpectedColumnStats(newLogicalRDD.output))
-      assert(newStats === newExpectedStats)
-    }
-  }
-
   test("SPARK-39834: build the constraints for LogicalRDD based on origin constraints") {
     def buildExpectedConstraints(attrs: Seq[Attribute]): ExpressionSet = {
       val exprs = attrs.flatMap { attr =>
@@ -3298,50 +3222,6 @@ class DataFrameSuite extends QueryTest
 case class GroupByKey(a: Int, b: Int)
 
 case class Bar2(s: String)
-
-/**
- * This class is used for unit-testing. It's a logical plan whose output and stats are passed in.
- */
-case class OutputListAwareStatsTestPlan(
-    outputList: Seq[Attribute],
-    rowCount: BigInt,
-    size: Option[BigInt] = None) extends LeafNode with MultiInstanceRelation {
-  override def output: Seq[Attribute] = outputList
-  override def computeStats(): Statistics = {
-    val columnInfo = outputList.map { attr =>
-      attr.dataType match {
-        case BooleanType =>
-          attr -> ColumnStat(
-            distinctCount = Some(2),
-            min = Some(false),
-            max = Some(true),
-            nullCount = Some(0),
-            avgLen = Some(1),
-            maxLen = Some(1))
-
-        case ByteType =>
-          attr -> ColumnStat(
-            distinctCount = Some(2),
-            min = Some(1),
-            max = Some(2),
-            nullCount = Some(0),
-            avgLen = Some(1),
-            maxLen = Some(1))
-
-        case _ =>
-          attr -> ColumnStat()
-      }
-    }
-    val attrStats = AttributeMap(columnInfo)
-
-    Statistics(
-      // If sizeInBytes is useless in testing, we just use a fake value
-      sizeInBytes = size.getOrElse(Int.MaxValue),
-      rowCount = Some(rowCount),
-      attributeStats = attrStats)
-  }
-  override def newInstance(): LogicalPlan = copy(outputList = outputList.map(_.newInstance()))
-}
 
 /**
  * This class is used for unit-testing. It's a logical plan whose output is passed in.
